feat: 함수 분리처리 - calculateCart
